[Kernel utils & syscalls]: moved SyscallTable and SyscallFFI to crate::utils::syscalls from crate::syscalls
