+++ kernel/src/utils/mod.rs
@@ -7,6 +7,7 @@ pub mod elf;
 pub mod errors;
 pub mod io;
 pub mod path;
+pub mod syscalls;
 pub mod ustar;
 
 use core::{
